requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.5
pytz==2020.1
sqlparse==0.4.4
tzdata==2023.3
